% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permDep-pkg.R
\docType{package}
\name{permDep-package}
\alias{permDep-package}
\alias{_PACKAGE}
\alias{permDep-packages}
\title{Permutation Test for General Dependent Truncation}
\description{
Implementations of permutation approaches to test for quasi-independence.
The implemented approaches are powerful against non-monotone alternatives and
thereby offer protection against erroneous assumptions of quasi-independence.
The proposed tests use either a conditional or a unconditional method to
evaluate the permutation p-value.
Users also have an option to carry out the proposed permutation tests in a
parallel computing fashion.
}
\references{
Chiou, S.H., Qian, J., and Betensky, R.A. (2017).
Permutation Test for General Dependent Truncation. \emph{Techreport}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/stc04003/permDep}
  \item Report bugs at \url{http://github.com/stc04003/permDep/issues}
}

}
\author{
\strong{Maintainer}: Sy Han (Steven) Chiou \email{schiou@utdallas.edu}

}
