\name{permDep-package}
\alias{permDep-package}
\docType{package}
\title{
  Permutation Test for General Dependent Truncation
}
\description{
  Implementations of permutation approaches to test for quasi-independence. The implemented approaches are powerful against non-monotone alternatives and thereby offer protection against erroneous assumptions of quasi-independence. The proposed tests use either a conditional or a unconditional method to evaluate the permutation p-value. Users also have an option to carry out the proposed permutation tests in a parallel computing fashion. 
}
\details{
  \tabular{ll}{
    Package: permDep
    Type: Package
    Version: 1.0-0
    Date: 2017-04-27
    License: GPL(>=3)
    LazyLoad: yes
  }
}
\author{
  Sy Han Chiou \email{schiou@hsph.harvard.edu}
}
\references{
  Chiou, S.H., Qian, J., and Betensky, R.A. (2017).
  Permutation Test for General Dependent Truncation
  \emph{Techreport} \bold{1}{1}: 1--10.
}
\keyword{ permDep }
